Ugrade, people, upgrade. 4.0 is here.

+++ modules/purge.py
@@ -32,15 +32,16 @@ async def del_msg(_, message: Message):
 @with_reply
 async def purge(client: Client, message: Message):
     chunk = []
-    async for msg in client.iter_history(
+    async for msg in client.get_chat_history(
         chat_id=message.chat.id,
-        offset_id=message.reply_to_message.message_id,
-        reverse=True,
+        limit=message.id - message.reply_to_message.id + 1,
     ):
-        chunk.append(msg.message_id)
+        if msg.id < message.reply_to_message.id:
+            break
+        chunk.append(msg.id)
         if len(chunk) >= 100:
             await client.delete_messages(message.chat.id, chunk)
-            chunk = []
+            chunk.clear()
             await asyncio.sleep(1)
 
     if len(chunk) > 0:
